deploy/app/requirements.txt
flask
jinja2
werkzeug
itsdangerous
PyJWT==2.3
cryptography
